Moved btn start methods to Main
